server player counts, names and server state to DB
This is still not done pefectly. A .LOCK file is populated with JSON data when `playerCounts.py` finishes executing. Whenever a request is made, a check is made to see if the .LOCK file has not been edited for TWO minutes. If this condition is true, the script will run again.

When the script runs, it causes the response time for the page to be roughly 300-900ms. This is not ideal at all ,even though it will only run at minimum every two minutes.

My thoughts are that I will need to explore spawning a new thread to run the data, however the models of the page still won't update if the page is loaded before the database has the new data. Likely still more to learn.