All the Mods 7 to website
In the future, I would like to design this such that simply adding a server to the SERVER_DATA setting, will fully create all needed functionality. At that point one will only need to add the images/information for the added server. As of now, the actual server query process is entirely as so (besides resetting the internal dictionary)