bot.py rewrite for slash commands
With discord.py (as far as I could tell), it is not possible to use Discord's new (and forced) slash command system when using a class based bot. Rewrote the bot to use decorators to define event listeners and now, commands.

The new command will accept an argument for what server key (ftbu, e6e, ct2, etc) to use. It will then craft an embed using data gathered from the server pertaining to the supplied key and send it in the channel the command was invoked from.

The embed contains;
- Modpack name
- Modpack address
- Modpack and server description
- Links to rules, banned items, and voting for specific server
- modpack version
- reminder to read stuff at the spawn
- the corresponding image for the modpack from the website

This is in a fuly functional state currently, I will certainly make some quality changes later.