Modify bot message to send two separate embeds
The first embed will contain nothing but the image, and will be at the top.

The second embed will contain all information. Detaching the image from the embed allows it to take up the maximum embed width, for readability. Placing the image at the top also looks a lot better