Site login form/page/template to Dropdown
General descript: The /login page and template is now loaded via ajax into the non-logged-in user's dropdown menu. Any form errors will be placed within the same dropdown if login credentials are incorrect (separate messages for non-existent user, and wrong password)

- Added clean() method to User Login form, checking if a user exists and if a password is incorrect, and raising corresponding ValidationErrors if so
- Modified URL and View name of User_Login to User_Login_Form
- Modified User_Login_Form to not get information that is not needed in the now AJAX/HTMX requested template, and modified redirection URL to place back at URL the user was at when submitting. Applied same redirect logic to logout function
- Removed check if a user is active
- Utilize Django Messages to send Form Validation Errors to the redirected URL
- Added CSS to modify the Bootstrap Form submit button, forcing it to 100% width. This is done here because the form is dynamically generated using a templatetag, rather than a direct form button element
- Moved Login with Discord button from base.html to the login.html AJAX template

One thing to note, it is absolutely possible for a user to directly access the AJAX templates by just entering the url "/login". This will take them to a rather incomplete looking page, as it is injected into the base template. I would like to find a way to prevent this, but I have simply loaded required CSS into the template so that it looks decent in the event that someone does access it.