Users/Discord Auth to own app authprofiles
Moved all functionality relating to User/oauth2/profiles to a new app 'authprofiles'

There is still a good amount of coupling between raptormc and authprofiles, I wanted to get the bulk of the code moved over before fine tuning.

Also wiped/remade migrations for raptormc. This was needed, as they insisted on refericing the DiscordAuthManager that was previously within raptormc.