Implement part of #58
DO_SERVER_QUERY setting attribute will now be checked the same as "in_maintenance" and "server_address" attributes. Servers will still be added to page, they will just not have their address queried.

Working on disabling the Player Counts on the front end when this parameter is set to False as well.