Implement basic control panel + fix bot start/stop
Control panel is now underway. Buttons exist for starting/stopping the bot, and import/exporting servers. Currently, only the bot start and stop is functional

The control panel template is extended from django-jazzmin's "base_site.html" template, which represents the admin page except for the content in the center. This integrates quite seamlessly with the existing admin.

Pressing start will use abstracted botware.get_bot_status to check that the bot is not running, then call abstracted botware.start_bot_process(). If the bot is running, a notifcation area will be populated telling you so.

Pressing stop will use the same check to make sure that the bot IS running, then call botware.stop_bot_process(). The same notifcation area will let you know if the bot is not running.

These actions are initiated via POST to new views, which will confirm that the user in the request has the "is_staff" attribute set to True before doing anything. GET requests to these views are redirected to the URL root.

Work will continue on implementing imports/exports and making the interface prettier.