Add creatable notification toasts
These toasts can be created in the admin, given a name and message. They can be enabled/disabled

These notifications will appear on the bottom right of the viewport with a sticky position.

They will only appear once per session. A session parameter "seen_notifications" is set to True in raptormc context processor on each request. The notifications will not display if this session parameter exists and is equal to True in a request.

These exist to give announcements that are specific to the website.