Page object for creating new site pages
Pages can now be created in the admin. With the following attributes:

name: the name of the page
content: the content of the page
gameservers_enabled: whether server buttons will appear on the page.

Pages can be linked to a NavigationLink, so that the Link will take you to the page. A NavigationLink's URL field will be ignored if it is linked to a Page