Working session checks for each notifications
This is probably extremely sloppy. Incrementing the session value to 4 seems to work quite well. By the time one actually gets to the page, the value will be 2 and the notification will show. After the second reload, the value will be 4 and the notification will not show. This is tracked for EACH TOAST, so if a new one is made that will appear one for each session.

Also added jazzmin icon for Page model in the admin