Add receiver to create ico when avatar changes
When an icon file is saved, the hash of the avatar image will be used as the filename. Every time a SiteInformation is saved while it has a avatar image, the hash of the current avatar will be checked against the filename. Only if the hashes do not match will the icon be created again. This will prevent a new icon from generating every time a save is made, or if the same image is re-uploaded.

Also added imports for previous commits