Redirect views to 404 if page is disabled
Within staffapps, this is done in a de-coupled fashion. With a try/except block, we're checking the DefaultPages model for boolean info on the attributes. If those checks fail to ModuleNotFound, they'll simply be skipped and the logic of the view will be accessed. If the boolean value is true and there is no error, the logic of the view will be accessed as well. This allows staffapps to be used without raptormc while still allowing the views to be "disabled".