Add trailing slash URLS for main pages
I'm honestly not certain if this was working fine before, but django's APPEND_URL feature does not seem to be working well. This is related to the issue that was occuring with redirects in Chrome during gameservers API requests. In this case, I want users to be able to access these pages with or without /, so this will exist for now.