Content to home.html and use hx-preserve/re-eval
home.html contains the contents of the homepage, whereas base.html is moreso a skeleton for the SPA. As such, home.html HEAD will immediately overwrite base.html's HEAD. As such, apply most base information from base.html's HEAD that pertains to site content to home.html

Also, use hx-preserve on tags that wil not change. This tag tells HTMX to never replace these tags. Unless, the tag also has hx-head="re-eval" The presence of both of these tags creates the following condition: The tag will stay if the replacing HEAD does not contain the same tag. If the replacing HEAD does contain the same tag, it will be replaced.

Currently, this presents a condition where if a default element is replaced, such as when the og:image's default Avatar value is replaced with a user profile picture on the user page, then that default will not return when the user leaves the replacing page. The default og:image will only return if the user navigates to the home page, or refreshes the page. This is a non-issue, as og tags are only used when a page is initally fetched in discord messages, tweets, etc.