Use avatar og:image in all other templates
Previously, the og:image was placed in base.html to apply to all other templates. When using both hx-preserve and hx-head, this seems to always be replaced even when the tag is not in the replacing HEAD. Added the og:image for the default avatar to all pages except for user to rectify this. This also removes the previous issue of the default og:image not applying after leaving a user page, although this was benign.

I'm not certain there is a good way to use DRY principles with the HEAD in these templates, but changes may be made to move towards that.