Remove opengraph data from snippet HEAD
These values will be found in routes.py, and inserted to context.

This needs to be done as opengraph data is only read from the initial page source with Discord, etc. Any data applied from htmx requests is not read. As such, this data has to be found and inserted to the template server-side, before the first canonical request.

With this, the opengraph data will not update at all when changing pages via SPA links. This is benign, as opengraph data is onl used in canonical requests.