Correct JS injection with discord messages
With how this was, any html tags (including <script>) included in a discord message will be parsed directly as HTML. This allows staff members to perform Js injection attacks using messages. This was not available for general users, a user would need to assigned Staff Role to be able to add messages to the app.

This has been corrected by modifying the existing template filter for removing discord references to also include HTML tags. This will apply before the rest of the tags inject anchors into the messages.

Also included is formatting changes for gameservers announcement list. raptorbot templates coming after.