multi stage build for production image
Tinkering with this still. The goal is to have a smaller image for use in production deployments.

The best use case is deploying with Portainer from the repository. The project has been designed to utilize the "stack.env" and environment variables defined within portainer. The project will also read a "stack.env" file located in the same directory as the docker-compose-prod.yml file if not deploying with portainer.

Current wrinkle is still adding your own web server cert/key without just adding them to the raptor-conf volume.