Check for informative text when making routes
informative_text has been added as a class attribute to Route, and will be filled with the informative text object associated with a route. This is checked using the "name" attribute of the URLPattern, capitalizing it and appending " Information" to find an informative text. If it does exist, the route will be identifed as a route using an informative text.

Then, the check will use it properly when returning context. This provides a way to define which pages should be checked for informative texts.

I've also moved the getting of "main routes" to it's own sub-function. The main function body is only checking for the routes against the path after running the "get" sub-functions