Retain players in database during server query
Modify PlayerPoller to no longer clear players on every run. Instead, each online player on a server will be checked against the current Players in the database. If they exist, their current server, last online time, and online values will be changed and saved.

If a Player does not exist, a new one will be created with the above data.

As PlayerPoller queries all servers, a list of all online player names is saved. After all servers are queried, this list of online players is checked against the database. If any Players in the database's online value is True, but they are NOT in the online player list, they will be set to offline.

This way, all players that have joined a server are saved, along with their last online time, and their current online status.

This will be expanded upon further to incorporate statistics of players over time on each server.