Check for same date
Since firefox will not supply a timestamp, we need to apply a timestamp range for the entire day if the same day is supplied for start and end.

If either of the two contains a T, then that means it did not come from firefox and we don't need to add the timestamp.