Add href to hx-get anchors
hx-get will always take precedence with a left click, loading the page with ajax as expected. Adding the href to the anchor allows you to also open the link in a new tab with the given anchor, only if you middle-click or right-click > open in new tab.

This operates exactly how I would like now. All links on the main site have had an href added to it, derived in a similar way as the hx-get without the api url parameters.