Receiever to send discord bot command if server changes
Checking only certain fields, a receiver in gameservers will set the update_embeds field of DiscordBotTasks to True if those fields have changed.

Set  a boolean field on a SentEmbedMessage object indicating that the server info has been changed but the embed itself in discord has not been edited yet.