Deleted SentEmbedMessage queue for bot to delete
When SentEmbedmessages are deleted from the admin, an object containing that messages ID and channel ID are added to a string in messages_to_delete, following a CSV structure.

If deleted_a_message is True, it will be set to False and the function will not run, as the message will already be deleted from Discord if this is so.