Redirect www. to non-www
This is mainly for the purpose of having consistent sessions for the app, such that users will always be logged into the account they were logged into. Currently, sessions for www and non-www are not shared, and efforts to use https://docs.djangoproject.com/en/4.2/ref/settings/#csrf-cookie-domain to share these were not successfull.

This seems to be how the majority of public apps do this to begin with.