Implement basic Stripe checkout flow
Donation packages can now be bought. CompletedDonation model will track a payment once the redirect has been issued.

Currently, commands are not sent.

Whether a minecraft username form is supplied is determined from whether a user is logged in. Checks need to be implemented to see if a user's Minecraft_username is empty, and supply a form to enter a username if so.

the success page needs to be overhauled. A route for the raptormc app needs to be made, and we need to implement some better checks to prevent those that have not donated recently from getting to that page, despite it not actually initiating any code.

A webhook endpoint for stripe to send payment results to will confirm if a payment is successful, and update the  package if so. At this point, commands would be run from the relevant package.