Allow discord username input on donation and use existing if it exists
Since we want to gather the Discord username from logged in users without discord-made accounts, all requests to this endpoint will now be via POST.

Discord usernames for donations are optional in all cases, except for when a user is logged into a discord-made account.

if a user is logged into a discord-made account, they will not need to enter anything to donate, all of the data will be pulled from their account

If a non-discord user is logged in, a form will be provided for them to enter a discord username , while pulling minecraft username from their account

If a user is not logged in at all, they will be given forms for both minecraft and discord username.