Check if package has roles or commands before giving or sending
In all instances where a completed donation will be issued roles/commands, check if the respective manytomany field has at least 1 object before running.

Error messages will be provided in the panel if trying to re-send for these donations. If the package is later updated to have roles or commands, you will be able to re-send them.