paypal module in payments
This flow utilizes standard mechanisms from django-paypal. It creates a compelted donation marked as incomplete the same as the stripe flow, however we are returning the django-paybal "Buy it Now" form. We are not going to be using this in a standard way in the view

This begs the question of why I did not just use Paypals API. This works is the answer.