IPN reciever listener
This code is called when paypal sends the POST request on successful payment, and django-paypal verifies it as valid. We are checking if the invoice ID matches what we have in the database, as well as the package price and receiver email.

We then complete the donation in the same way it is handled via Stripe.