paypal redirect template
This template renders the "Buy it Now" django-paypal button as hidden, and displays a simple message that you are being redirected.

A script exists here, that will automatically submit the form. This is a simple way that I found to automatically direct the donator to the paypal payment page, without having separate buttons for Paypal/stripe checkout and adding a second click in between.

As far as I can tell, there is not a downside to doing it this way. If a user has Javascript disabled, they're fucked anyway.