Server update CRUD view
This extends UpdateView, overrding post() to return the form messages to the SPA without reloading the page. This generic view can be used for any other models that will be given a CRUD interface here

Models extending this can supply an "ignored_fields" list to give a list of fields to ignore for validation and updating of models. Can also provide image_fields to update image fields properly

For servers, a custom form is in use to use TinyMCE text editor. There is also a custom widget for picture fields to display the uploaded image alongside the form to upload.

Further, changed all success messages to replace underscores in the changed string with spaces