form control for server
Writing some fun js to toggle button states. These will be used to hide and show parts of the server form

Currently the buttons are just changing their color when clicked to primary, and others to secondary. functions wired up to toggle collapses or some other element to show and hide sections