Detach panel from django admin
panel_base.html is no longer extending the django admin.

Work in the head tag and head merging from other compnents is not completed

This acts as an SPA very similar to the main app.

The goal being to have CRUD interfaces for all models available here in an organized fashion

a top bar is loaded very similar to main app, with slightly less data. The profile dropdown here will show a link back to site instead of to the panel, and the link to the profile is not  shown.

The main branding image here links to the panel homepage