Creatable staff applications
This is one large commit.

New Staff Applications can be created, with a name and a list of Form Fields attached, also creatable. Each staff application is loaded on the site.

Submitted Staff Applications are made with the POST, after converting to JSON and saving to a JSON field. When loading for read views, this is parsed to a a viewable form.

Implements https://github.com/zediious/raptor-web/issues/96