Previous PR append - wait for databse
The real cause the problem here is that the database migration was failing due to the database not being connectable on startup. The statement is just skipped, so the databse is never migrated as the database is up by the time the app is starting.

We'll wait 15 seconds in the startup script before running the migration, ensuring that the database has started. Looking at logs, the database is ready for connections a few seconds after the error is thrown so this should be sufficient.