Ensure bulk delete stays disabled through filtering
Using any sort, search, or pagination would cause alpine state to get weird and show button enabled until you check a box.

On list load, set the class disabled. Any click will render the button usable again with a new click event on the checkboxes to remove the disabled class from the bulk delete button

This is needed due to the page swapping out the entire page. Not sure what is actually getting lost as selected is empty and the bool is false when the issue was happening.

This resolves the issue nonetheless. Also add