Toggle panel sidebar with ALT+A keyboard shortcut
Utilize Alpine to implement an ALT+A keyboard shortcut to open the offcanvas nav. This can be done from anywhere in the panel, even when scrolled all the way to the bottom/away from top bar.

We're instantiating a bootstrap offcanvas based on the panel offcanvas ID, and storing that. This is then toggled when the shortcut is used.