Timeout minecraft server queries very quickly
This timeout is set for quite some time by default. This can cause 6 servers who are offline to take over a minute to show as so, with the request waiting the whole time. With 0.5 set, a failed server takes about 4 seconds to return, so this will be reduced overall. This does beg the notion of writing this custom, as it would allow us to fine tune that and also support other game servers as is desired in the future.