Use hx-oob to swap server button and player count info
Now, the player and server button queries return only the small amount of data within the buttons and modals, and are swapping them in using hx-oob.

The modals and buttons are now fully loaded into the loading template, and only the small bits inside are swapped.

This has two boons;
1) The serverbox area no longer adjusts height when query is made, preventing the page being jarred.
2) You can leave the player name modal open durinng a query and it will update the names inside, and will no longer force a refresh.

The query refresh process is now very smooth.