Apply routes to routable models
This is the way that models will be given canonical routes going forward.

A "routes" method will return a tuple with a list of tuples. Each of these represents a route to be added. Each tuple will contain at minimum one nested tuple. More than one can be passed. The first will contain the canonical/user-accessed URL for this model. The second (optional) item in the tuple can contain an expression. IF this evaluates to True, the route will be added. If false, it will not.

A "route_name" method will also return a readable type for this route, to be used if wanting to pass specific context to templates based on the model (currently used for opengraph data in main app routes)

This has been applied to all currently routeable models.