middleware to reroute without hx-request
This middleware will check for the HX-Request header to be true

It will also check that `/api/` is within the request path, as this will always be the case for requests made by htmx. If it's missing, the response will be returned.

If both HX-Request is NOT true and `/api/` is present in the request path, the request will be redirected to `/404`