change user filter request trigger from click to change
Previously, along with keyup, the user list filter would trigger on any click on the whole form element, which was quite large.

Now, it will only trigger on change to the checkbox element, and keyup after a delay. This makes this element much smoother to use.

We also remove a useless submit button, this was relevant in an earlier iteration.